requirements.txt
pandas
dash==2.15.0
plotly-express==0.4.1
dash-bootstrap-components==1.5.0
gunicorn